chore(SAV-106): add a test

+++ src/Modules/Transactions/Domain/Finance/Money.cs
@@ -2,11 +2,11 @@
 
 public record Money
 {
-    public required decimal Amount { get; init; }
+    public required int Amount { get; init; }
     public required Currency Currency { get; init; }
 
-    public static Money From(decimal amount, Currency currency) => new() { Amount = amount, Currency = currency };
-    public static Money From(decimal amount, string currency) => new() { Amount = amount, Currency = Currency.From(currency) };
+    public static Money From(int amount, Currency currency) => new() { Amount = amount, Currency = currency };
+    public static Money From(int amount, string currency) => new() { Amount = amount, Currency = Currency.From(currency) };
 
     private Money()
     { }
